Make server and internet clone models more UI independent
- move ServerSettingsModel to LibChorus
- add new InternetCloneSettingsModel in LibChorus that has most of
  the previous functionality of GetCloneFromInternetModel but is
  strictly UI independent (it has a progress object, but that is
  UI agnostic)
- derive GetCloneFromInternetModel from InternetCloneSettingsModel
  and wrap a few methods that catch errors and display UI to the
  user.
