+++ src/Chorus/UI/Clone/GetCloneFromInternetModel.cs
@@ -1,103 +1,28 @@
 ﻿using System;
-using System.IO;
 using System.Media;
 using System.Threading;
-using Chorus.UI.Misc;
-using Chorus.Utilities;
-using Chorus.VcsDrivers;
-using Chorus.VcsDrivers.Mercurial;
 using L10NSharp;
 using SIL.Progress;
 using SIL.Reporting;
+using Chorus.Model;
+using Chorus.Utilities;
+using Chorus.VcsDrivers.Mercurial;
 
 namespace Chorus.UI.Clone
 {
-	public class GetCloneFromInternetModel : ServerSettingsModel
+	public class GetCloneFromInternetModel : InternetCloneSettingsModel
 	{
 		private bool _showCloneSpecificSettings;
-		private MultiProgress _progress;
 
-
-		public GetCloneFromInternetModel(string parentDirectoryToPutCloneIn)
-		{
-			_showCloneSpecificSettings = true;
-			ParentDirectoryToPutCloneIn = parentDirectoryToPutCloneIn;
-			_progress = new MultiProgress();
-		}
-		public GetCloneFromInternetModel()
+		public GetCloneFromInternetModel(): base()
 		{
 			_showCloneSpecificSettings = false;
-			_progress = new MultiProgress();
 		}
 
-		public string ParentDirectoryToPutCloneIn { get; set; }
-
-
-		public override void InitFromUri(string url)
+		public GetCloneFromInternetModel(string parentDirectoryToPutCloneIn): base(parentDirectoryToPutCloneIn)
 		{
-			base.InitFromUri(url);
-			LocalFolderName = UrlHelper.GetValueFromQueryStringOfRef(url, @"localFolder", string.Empty);
+			_showCloneSpecificSettings = true;
 		}
-
-		public bool ReadyToDownload
-		{
-			get
-			{
-				return HaveNeededAccountInfo && HaveWellFormedTargetLocation && TargetLocationIsUnused;
-			}
-		}
-
-		public bool TargetLocationIsUnused
-		{
-			get
-			{
-				try
-				{
-					// the target location is "unused" if either the Target Destination doesn't exist OR
-					//  if it has nothing in it (I tried Clone once and it failed because the repo spec
-					//  was wrong, but since it had created the Target Destination folder, it wouldn't
-					//  try again-rde)
-					return Directory.Exists(ParentDirectoryToPutCloneIn) &&
-						   (!Directory.Exists(TargetDestination)
-						   || (Directory.GetFiles(TargetDestination, "*.*", SearchOption.AllDirectories).Length == 0));
-				}
-				catch (Exception)
-				{
-					return false;
-				}
-			}
-		}
-
-		public bool HaveWellFormedTargetLocation
-		{
-			get
-			{
-				return (!string.IsNullOrEmpty(LocalFolderName) && LocalFolderName.LastIndexOfAny(Path.GetInvalidPathChars()) == -1);
-			}
-		}
-
-		public string LocalFolderName { get; set; }
-
-		public string TargetDestination
-		{
-			get
-			{
-				if(string.IsNullOrEmpty(LocalFolderName))
-				{
-					return "";
-				}
-				return Path.Combine(ParentDirectoryToPutCloneIn, LocalFolderName);
-			}
-		}
-
-
-		public bool TargetHasProblem
-		{
-			get {
-				return !HaveWellFormedTargetLocation || !TargetLocationIsUnused;
-			}
-		}
-
 
 		public bool ShowCloneOnlyControls
 		{
@@ -125,25 +50,11 @@ namespace Chorus.UI.Clone
 		///<summary>
 		///</summary>
 		///<returns>true of successful; false if failed</returns>
-		public bool SetRepositoryAddress()
+		public override bool SetRepositoryAddress()
 		{
 			try
 			{
-				var repo = new HgRepository(TargetDestination, _progress);
-				var name = new Uri(URL).Host;
-				if (String.IsNullOrEmpty(name)) //This happens for repos on the local machine
-				{
-					name = @"LocalRepository";
-				}
-				if (name.ToLower().Contains(@"languagedepot"))
-					name = @"LanguageDepot";
-
-				var address = RepositoryAddress.Create(name, URL);
-
-				//this will also remove the "default" one that hg puts in, which we don't really want.
-				repo.SetKnownRepositoryAddresses(new[] { address });
-				repo.SetIsOneDefaultSyncAddresses(address, true);
-				return true;
+				return base.SetRepositoryAddress();
 			}
 			catch (Exception error)
 			{
@@ -152,20 +63,16 @@ namespace Chorus.UI.Clone
 			}
 		}
 
-		public void DoClone()
+		public override void DoClone()
 		{
 			try
 			{
-				//review: do we need to get these out of the DoWorkEventArgs instead?
-				var actualCloneLocation = HgRepository.Clone(new HttpRepositoryPath(URL, URL, false), TargetDestination, _progress);
-				LocalFolderName = Path.GetFileName(actualCloneLocation.TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar));
+				base.DoClone();
 				using (SoundPlayer player = new SoundPlayer(Properties.Resources.finishedSound))
 				{
 					player.PlaySync();
 				}
-
 			}
-
 			catch (Exception error)
 			{
 				using (SoundPlayer player = new SoundPlayer(Properties.Resources.errorSound))
@@ -198,29 +105,6 @@ namespace Chorus.UI.Clone
 		public void Click_FixSettingsButton()
 		{
 			_progress.CancelRequested = false;
-		}
-
-		public void CleanUpAfterErrorOrCancel()
-		{
-			if (Directory.Exists(TargetDestination))
-			{
-				Directory.Delete(TargetDestination, true);
-			}
-		}
-
-		public void AddMessageProgress(IProgress p)
-		{
-			_progress.AddMessageProgress(p);
-		}
-
-		public void AddStatusProgress(IProgress p)
-		{
-			_progress.AddStatusProgress(p);
-		}
-
-		public void AddProgress(IProgress p)
-		{
-			_progress.Add(p);
 		}
 	}
 }
