+++ src/LibChorus/LibChorus.csproj
@@ -379,6 +379,8 @@
     <Compile Include="VcsDrivers\Mercurial\Revision.cs" />
     <Compile Include="VcsDrivers\Mercurial\HgRunner.cs" />
     <Compile Include="Utilities\HgProcessOutputReader.cs" />
+    <Compile Include="Model\ServerSettingsModel.cs" />
+    <Compile Include="Model\InternetCloneSettingsModel.cs" />
   </ItemGroup>
   <ItemGroup>
     <BootstrapperPackage Include="Microsoft.Net.Framework.2.0">
@@ -481,4 +483,7 @@
     <PreBuildEvent>
     </PreBuildEvent>
   </PropertyGroup>
+  <ItemGroup>
+    <Folder Include="Model\" />
+  </ItemGroup>
 </Project>
