+++ src/LibChorusTests/Model/InternetCloneSettingsModelTests.cs
@@ -1,19 +1,19 @@
 ﻿using System.IO;
-using Chorus.UI.Clone;
 using NUnit.Framework;
 using SIL.TestUtilities;
+using Chorus.Model;
 
-namespace Chorus.Tests.UI.Clone
+namespace LibChorus.Tests.Model
 {
 	[TestFixture]
-	public class GetCloneFromInternetModelTest
+	public class InternetCloneSettingsModelTests
 	{
 		[Test]
 		public void InitFromUri_GivenCompleteUri_AllPropertiesCorrect()
 		{
 			using (var testFolder = new TemporaryFolder("clonetest"))
 			{
-				var model = new GetCloneFromInternetModel(testFolder.Path);
+				var model = new InternetCloneSettingsModel(testFolder.Path);
 				model.InitFromUri("http://john:myPassword@hg-languagedepot.org/tpi?localFolder=tokPisin");
 				Assert.AreEqual("tokPisin", model.LocalFolderName);
 				Assert.IsTrue(model.ReadyToDownload);
@@ -26,7 +26,7 @@ namespace Chorus.Tests.UI.Clone
 		{
 			using (var testFolder = new TemporaryFolder("clonetest"))
 			{
-				var model = new GetCloneFromInternetModel(testFolder.Path);
+				var model = new InternetCloneSettingsModel(testFolder.Path);
 				model.AccountName = "account";
 				model.Password = "password";
 				model.ProjectId = "id";
@@ -39,7 +39,7 @@ namespace Chorus.Tests.UI.Clone
 		{
 			using (var testFolder = new TemporaryFolder("clonetest"))
 			{
-				var model = new GetCloneFromInternetModel(testFolder.Path);
+				var model = new InternetCloneSettingsModel(testFolder.Path);
 				model.LocalFolderName = "SomeFolder";
 				// Ideally would call model to start the clone - but that's in the dialog for now so fake it instead.
 				Directory.CreateDirectory(model.TargetDestination);
